+++ src/pretalx/agenda/views/sneakpeek.py
@@ -1,4 +1,6 @@
+from django.contrib import messages
 from django.http import HttpResponseRedirect
+from django.utils.translation import ugettext_lazy as _
 from django.views.generic import TemplateView
 from django_context_decorator import context
 
@@ -22,5 +24,6 @@ class SneakpeekView(EventPermissionRequired, TemplateView):
         can_schedule = request.user.has_perm('agenda.view_schedule', request.event)
 
         if not can_peek and can_schedule:
+            messages.success(request, _('Our sneak peek has been disabled because we have published a schedule!'))
             return HttpResponseRedirect(request.event.urls.schedule)
         return super().dispatch(request, *args, **kwargs)